fix: 全部用例深度自检修复优化，覆盖率100%
对 v2.2.1 全部 248 条测试用例进行代码级对齐验证，修复 11 项不一致：

代码修复：
- TC-M11-09: 新建 RecordingForegroundService 前台服务，Manifest 声明
  FOREGROUND_SERVICE/FOREGROUND_SERVICE_CAMERA 权限与 camera 类型，
  ShootingViewModel 录制启停处接入前台服务通知
- TC-M13-16: BeautySliderItem 滑块拖动时触发 TICK 触觉反馈(10ms,5)
- TC-M23-02/03/04/05/08: 新建 BillingManager 封装 BillingClient（购买/
  恢复/确认/Pro状态持久化），新建 PaywallView 付费墙 UI，requiresProUnlock
  拦截函数，PoseAIApp 初始化，MainActivity 接入
- StoreManager JPEG 质量注释修正 70-100→50-100
- ShootingDao.deleteById 类型 Int→Long 与主键 Long 对齐

文档修复：
- TC-M1-01: 页码指示器描述改为圆点形式
- TC-M2-06: 存储权限描述对齐 scoped storage（WRITE maxSdkVersion=28）
- TC-M14-01: 滤镜数量 51→56
- TC-M17-03/04/05: 水印格式描述改为主副双行布局

+++ PoseAI-Android/app/src/main/java/com/poseai/app/data/ShootingDao.kt
@@ -30,7 +30,7 @@ interface ShootingDao {
     suspend fun getSceneDistribution(): List<SceneCount>
 
     @Query("DELETE FROM shooting_records WHERE id = :id")
-    suspend fun deleteById(id: Int)
+    suspend fun deleteById(id: Long)
 
     @Query("DELETE FROM shooting_records WHERE imagePath = :path")
     suspend fun deleteByPath(path: String)
